// src/modules/core/auth/translate.enum.ts
export enum AuthKeys {
    USER_NOT_FOUND = "auth.USER_NOT_FOUND",
    ROLE_NOT_FOUND = "auth.ROLE_NOT_FOUND",
    EMAIL_EXISTS = "auth.EMAIL_EXISTS",
    EMAIL_NOT_EXISTS = "auth.EMAIL_NOT_EXISTS",
    VERIFICATION_CODE_NOT_VALID = "auth.VERIFICATION_CODE_NOT_VALID",
    VERIFICATION_CODE_EXPIRED = "auth.VERIFICATION_CODE_EXPIRED",
    WRONG_PASSWORD = "auth.WRONG_PASSWORD",
    EMAIL_VERIFICATION_CODE_SENT_SUCCESSFULLY = "auth.EMAIL_VERIFICATION_CODE_SENT_SUCCESSFULLY",
    EMAIL_RESET_PASSWORD_CODE_SENT_SUCCESSFULLY = "auth.EMAIL_RESET_PASSWORD_CODE_SENT_SUCCESSFULLY",
    RESET_PASSWORD_CODE_VERIFIED_SUCCESSFULLY = "auth.RESET_PASSWORD_CODE_VERIFIED_SUCCESSFULLY",
    RESET_PASSWORD_SUCCESSFULLY = "auth.RESET_PASSWORD_SUCCESSFULLY",
    INVALID_REFRESH_TOKEN = "auth.INVALID_REFRESH_TOKEN",
}